Fix: response lost
I am not sure where to start, we run against ory platform and noticed
that the response "from" oathkeeper did not contain a subject when it
got to the backend service.

I managed to find out that the "body" in `forwardRequestToSessionStore()`
was nil/empty in some cases and the gjson calls failed silently.

I started log.Printf() debugging in these two service files and simplified
the code a bit to make it more readable. And that seemed to have fixed it.